Move tests to testsuite and add Jayway JsonPath compatibility tests
