+++ testsuite/pom.xml
@@ -11,7 +11,9 @@
     <packaging>pom</packaging>
 
     <modules>
-        <module>docker</module>
+        <module>custom-check-test</module>
+
+        <!--module>docker</module>
         <module>client-secret-jwt-keycloak-test</module>
         <module>access-token-introspection-keycloak-test</module>
         <module>refresh-token-jwt-keycloak-test</module>
@@ -22,7 +24,7 @@
         <module>jwt-over-plain-keycloak-authz-test</module>
         <module>introspection-over-plain-keycloak-authz-test</module>
         <module>client-secret-jwt-keycloak-audience-test</module>
-        <module>client-secret-introspection-keycloak-audience-test</module>
+        <module>client-secret-introspection-keycloak-audience-test</module-->
     </modules>
 
     <licenses>
